refactor: Move create_app to factory and update CI
This commit refactors the application by moving the `create_app` function into a dedicated `app/factory.py` module. This improves the project structure and follows best practices for Flask application development.

Additionally, the CI workflow in `.github/workflows/ci.yml` has been updated to:
- Use Python 3.12.
- Install the project in editable mode (`pip install -e .`) to ensure tests run correctly.
- Implement a proper orphan-branch flow for publishing the coverage badge.
- Remove the problematic `git checkout -` step.

These changes address the CI failures and improve the overall maintainability of the project.
