+++ .github/workflows/ci.yml
@@ -9,7 +9,7 @@ on:
     branches: [ "main" ]
 
 permissions:
-  contents: write   # allow pushing the badge
+  contents: write
 
 jobs:
   test:
@@ -20,19 +20,19 @@ jobs:
       - name: Checkout code
         uses: actions/checkout@v4
         with:
-          # fetch badges branch too (if it exists)
           fetch-depth: 0
 
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
-          python-version: '3.11'
+          python-version: '3.12'
           cache: 'pip'
 
-      - name: Install deps
+      - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+          [ -f requirements.txt ] && pip install -r requirements.txt || true
+          pip install -e .[test]
           pip install -r tests/requirements.txt
 
       - name: Run tests with coverage
@@ -45,30 +45,9 @@ jobs:
 
       - name: Publish badge to 'badges' branch
         run: |
-          # Configure git author for CI
           git config user.name "github-actions[bot]"
           git config user.email "41898282+github-actions[bot]@users.noreply.github.com"
-
-          # Switch to (or create) badges branch
-          git fetch origin badges || true
-          if git show-ref --verify --quiet refs/heads/badges; then
-            git checkout badges
-          else
-            git checkout -b badges
-          fi
-
-          # Move the new badge into this branch root
-          mv -f coverage.svg coverage.svg
-
-          # Commit only if changed
-          git add coverage.svg
-          if ! git diff --cached --quiet; then
-            git commit -m "chore: update coverage badge"
-            git push origin badges
-          else
-            echo "No badge changes to publish."
-          fi
-
-      - name: Back to main (no-op)
-        if: always()
-        run: git checkout -
+          git checkout --orphan badges-temp
+          git add -f coverage.svg
+          git commit -m "chore: update coverage badge"
+          git push -f origin HEAD:badges
